requirements.txt
flask
pyyaml
google-cloud-logging
google-cloud-firestore
google-cloud-aiplatform
langchain-google-vertexai
langchain-community
langchain-experimental==0.3.4
pymupdf
# Jinja2
# pytest
# firebase-admin
# requests==2.32.3
# protobuf==5.29.5
# langchain
# langchain-core<2.0.0
# langchain-text-splitters<2.0.0